Split punch_holes function in several smaller ones
punch_holes: the main one, it calls the others and the result is a
sudoku with some holes in it. The one which will get presented to the
user.

get_random_punch_position: randomly generates a number at which position
we want to punch, it also keeps track of where we punched already using
is_fresh_number.

is_fresh_number: remembers all the places we punched.

punch_hole_maintain_unique: punches the whole and tries to solve it, to
find out if its unique.