Provide hints based on user_board
hint() starts to give hints from the user_board now, so taking in
account everything the user already filled in. At first I wanted to take
hints only from plain_board to assure that there will be a valid hint.
But maybe it makes more sense like this.

Also, now we first check if the board is solvable and if it is we try to
solve it.

We only show "hint provided" if it's solvable and we actually provided
an hint. So no message means the user needs to fix his board first. So
he knows there is something wrong with it. Maybe we should tell him?

VERSION get's set by autotools, but it's possible that someone wants to
include it incase it needs it for debugging. So the comment should be at
the right point to make it quicker.