Fix array access in validitycheck
Changed coding style to match the rest.
There was a problem that we went through the numbers 1 to 9 in ASCII 49
to 57 and made minus 48 to get the real numbers. This was then used as
indices to the array. But the array is from 0 to 8, so we had to do a
minus 49 instead of 48.

Should fix issue #4.