Only update grind upon movement if still playing
Mistake reproducer:
Press 'S' to solve the puzzle. Use the movment keys.
Effect:
The puzzle was still 'finished' but the grid showed like it was before
solving.

Mistake was introduced with 710a5f6fb0535f49d02a7fa451cc53f06ad1b675

If we have highlighting enabled, we need to redraw the whole grid
so we can have the new colors in the matching colors.
this should only be done when we are playing, because plain_board
is actually the one being solved and thus displayed.
this is true for all movement keys.